Updated the crawl_files and added the is_url_valid
Given the file name as the parameter to read_files. Aslo . change the position of the bookkeeping.json to the place outside the WEBPAGES folder to the one in which the program is . So it will be in the folder of the program and inside we will have the WEBPAGES folder.